Moved rcVector benchmarks to their own file
